Refactoring errors and adding codes/types provided by Stripe + go fmt
