+++ charge.go
@@ -54,7 +54,7 @@ type ChargeParams struct {
 	Amount int64
 
 	// 3-letter ISO code for currency. Refer to the Stripe docs for currently
-    // supported currencies: https://support.stripe.com/questions/which-currencies-does-stripe-support
+	// supported currencies: https://support.stripe.com/questions/which-currencies-does-stripe-support
 	Currency string
 
 	// (Optional) Either customer or card is required, but not both The ID of an
