+++ coupon_test.go
@@ -135,7 +135,7 @@ func TestListCoupon(t *testing.T) {
 		t.Errorf("Expected Coupon List, got Error %s", err.Error())
 	}
 
-	// since we added 2 dummy coupons, we expect the array to be a size of 2 
+	// since we added 2 dummy coupons, we expect the array to be a size of 2
 	if len(coupons) != 2 {
 		t.Errorf("Expected 2 Coupons, got %s", len(coupons))
 	}
